moved generic object type to common
