+++ next-app/src/common/object-types.ts
